requirements.txt
absl-py==2.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
astunparse==1.6.3
atomicwrites==1.4.1
attrs==25.4.0
beautifulsoup4==4.14.2
blinker==1.9.0
blis==1.3.3
cachetools==6.2.4
catalogue==2.0.10
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.3.2
cycler==0.12.1
cymem==2.0.13
deepface==0.0.96
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
exceptiongroup==1.3.1
filelock==3.20.0
fire==0.7.1
Flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.12.19
fonttools==4.60.1
fsspec==2025.10.0
gast==0.4.0
gdown==5.2.0
google-auth==2.46.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.74.0
gunicorn==23.0.0
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
keras==2.13.1
Keras-Preprocessing==1.1.2
kiwisolver==1.4.9
libclang==18.1.1
lightecc==0.0.3
lightphe==0.0.15
lz4==4.4.5
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
mediapipe==0.10.31
ml_dtypes==0.5.4
mpmath==1.3.0
mtcnn==1.0.0
murmurhash==1.0.15
mysql-connector-python==9.5.0
namex==0.1.0
networkx==3.4.2
numpy==1.26.4
oauthlib==3.3.1
onnxruntime==1.23.2
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pluggy==1.6.0
plyer==2.1.0
polars==1.35.2
polars-runtime-32==1.35.2
preshed==3.0.12
protobuf==4.25.8
psutil==7.1.3
py==1.11.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyMySQL==1.1.2
pyparsing==3.2.5
pyreadline3==3.5.4
PySocks==1.7.1
pyTelegramBotAPI==4.33.0
pytest==7.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-telegram-bot==22.7
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
retina-face==0.0.17
rich==14.2.0
rsa==4.9.1
safetensors==0.7.0
schedule==1.2.2
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.2.0
shellingham==1.5.4
six==1.17.0
smart_open==7.5.1
sounddevice==0.5.3
soupsieve==2.8
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
sympy==1.14.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.3.0
thinc==8.3.10
threadpoolctl==3.6.0
tokenizers==0.22.2
tomli==2.3.0
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.3
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.231
ultralytics-thop==2.0.18
urllib3==2.6.2
wasabi==1.1.3
weasel==0.4.3
websockets==15.0.1
Werkzeug==3.1.4
wrapt==2.0.1
